Share privacy text redaction
