+++ scripts/entrypoints/run-e2e-smoke.sh
@@ -52,6 +52,7 @@ SWIFT_SOURCES=(
     "Sources/TranscriptedCore/Logging/AppLogger.swift"
     "Sources/TranscriptedCore/Logging/FileLogger.swift"
     "Sources/TranscriptedCore/Logging/LogTailTrimmer.swift"
+    "Sources/TranscriptedCore/Logging/PrivacyTextRedactor.swift"
     "Sources/TranscriptedCore/Logging/LogPrivacySanitizer.swift"
     "Sources/TranscriptedCore/Utilities/FilePermissions.swift"
     "Sources/UI/Shared/MeetingAudioArchiveResolver.swift"
